C# code using ToLowerInvariant (avg 26.4 ms)
Simple C# code using ToLowerInvariant or ToUpperInvariant average run time is 26.4 ms - ToLowerInvariant will disregard what language you're typing in.

For example if you want to lowercase "IÎÏ" into Turkish lowercase then using ToLower should respect that culture casing and return "ıiı" instead of "iîï".

That did not seem to work for me for unknown reasons and I have tried to trouble shoot that with CultureInfo.CurrentCulture = new CultureInfo("tr-TR"); and used Console.WriteLine(CultureInfo.CurrentCulture.Name); to make sure that Turkish was active as a current culture. If you have any ideas to not hesitate to commen in my blog post!

Eitherway, ToLowerInvariant is faster than ToLower as a run time which is helpful if you expect everything to be in the same language.