StrBuilder and conditional operator avg 21.8ms
Using the String Builder to Append the conversion into a new string. The conversion method is using the ?: (conditional operator) which is similar to if/then/else statements but can be a lot faster if your conditions are very simple to begin with.

This code is marginally faster than the Array Method but what if on a large scale the Array Method is going to be slower and will take too much memory? Worth trying it in a future test!